refactor: split setting viewmodel responsibilities
