+++ dogArea/Source/Domain/Profile/Models/SeasonProfileSummary.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct SeasonProfileSummary: Equatable {
+    let weekKey: String
+    let score: Int
+    let rankTier: SeasonRankTier
+    let contributionCount: Int
+}
