+++ dogArea/Views/ProfileSettingView/SettingViewModel.swift
@@ -1,20 +1,5 @@
-//
-//  SettingViewModel.swift
-//  dogArea
-//
-//  Created by 김태훈 on 11/13/23.
-//
-
 import Foundation
 import Combine
-import UIKit
-
-struct SeasonProfileSummary: Equatable {
-    let weekKey: String
-    let score: Int
-    let rankTier: SeasonRankTier
-    let contributionCount: Int
-}
 
 final class SettingViewModel: ObservableObject {
     enum ProfileEditValidationError: LocalizedError {
@@ -55,17 +40,19 @@ final class SettingViewModel: ObservableObject {
     @Published var seasonProfileSummary: SeasonProfileSummary? = nil
     @Published var isCaricatureGenerating: Bool = false
     @Published var isAccountDeletionInProgress: Bool = false
-    private let profileRepository: ProfileRepository
-    private let imageRepository: ProfileImageRepository
-    private let petManagementService: SettingsPetManaging
-    private let accountDeletionService: AccountDeletionServiceProtocol
-    private let authSessionStore: AuthSessionStoreProtocol
-    private let walkRepository: WalkRepositoryProtocol
-    private let featureFlags = FeatureFlagStore.shared
-    private let metricTracker = AppMetricTracker.shared
-    private let caricatureClient = CaricatureEdgeClient()
-    private var cancellables: Set<AnyCancellable> = []
-    private var uiTestPetManagementUserInfoOverride: UserInfo? = nil
+
+    let profileRepository: ProfileRepository
+    let imageRepository: ProfileImageRepository
+    let petManagementService: SettingsPetManaging
+    let accountDeletionService: AccountDeletionServiceProtocol
+    let authSessionStore: AuthSessionStoreProtocol
+    let walkRepository: WalkRepositoryProtocol
+    let seasonProfileSummaryService: SettingsSeasonProfileSummaryProviding
+    let featureFlags = FeatureFlagStore.shared
+    let metricTracker = AppMetricTracker.shared
+    let caricatureClient = CaricatureEdgeClient()
+    var cancellables: Set<AnyCancellable> = []
+    var uiTestPetManagementUserInfoOverride: UserInfo? = nil
 
     var pets: [PetInfo] {
         userInfo?.pet ?? []
@@ -79,17 +66,27 @@ final class SettingViewModel: ObservableObject {
         pets.filter { $0.isActive == false }
     }
 
-    private var isUITestPetManagementStubEnabled: Bool {
+    var isUITestPetManagementStubEnabled: Bool {
         ProcessInfo.processInfo.arguments.contains("-UITest.PetManagementStub")
     }
 
+    /// 설정 화면 뷰모델 의존성을 구성하고 초기 상태를 로드합니다.
+    /// - Parameters:
+    ///   - profileRepository: 사용자/반려견 프로필 읽기·저장을 담당하는 저장소입니다.
+    ///   - imageRepository: 프로필 이미지 업로드를 담당하는 저장소입니다.
+    ///   - petManagementService: 반려견 추가/수정/활성화 관리를 담당하는 서비스입니다.
+    ///   - accountDeletionService: 회원탈퇴 요청을 처리하는 서비스입니다.
+    ///   - authSessionStore: 인증 세션/식별자 조회를 담당하는 저장소입니다.
+    ///   - walkRepository: 다각형 영역 데이터를 읽는 저장소입니다.
+    ///   - seasonProfileSummaryService: 시즌 진행 현황 요약을 계산하는 서비스입니다.
     init(
         profileRepository: ProfileRepository = DefaultProfileRepository.shared,
         imageRepository: ProfileImageRepository = SupabaseProfileImageRepository.shared,
         petManagementService: SettingsPetManaging = SettingsPetManagementService(),
         accountDeletionService: AccountDeletionServiceProtocol = SupabaseAccountDeletionService.shared,
         authSessionStore: AuthSessionStoreProtocol = DefaultAuthSessionStore.shared,
-        walkRepository: WalkRepositoryProtocol = WalkRepositoryContainer.shared
+        walkRepository: WalkRepositoryProtocol = WalkRepositoryContainer.shared,
+        seasonProfileSummaryService: SettingsSeasonProfileSummaryProviding = SettingsSeasonProfileSummaryService()
     ) {
         self.profileRepository = profileRepository
         self.imageRepository = imageRepository
@@ -97,858 +94,25 @@ final class SettingViewModel: ObservableObject {
         self.accountDeletionService = accountDeletionService
         self.authSessionStore = authSessionStore
         self.walkRepository = walkRepository
+        self.seasonProfileSummaryService = seasonProfileSummaryService
         bindSelectedPetSync()
         bindAuthSessionSync()
         fetchModel()
         reloadUserInfo()
     }
+
+    /// 설정 화면에서 사용할 산책 영역 데이터를 로드합니다.
     func fetchModel() {
-        self.polygonList = self.walkRepository.fetchPolygons()
+        polygonList = walkRepository.fetchPolygons()
     }
 
+    /// 현재 사용자/선택 반려견/시즌 요약 상태를 한 번에 다시 로드합니다.
     func reloadUserInfo() {
-        self.userInfo = resolveDisplayedUserInfo()
-        self.selectedPet = profileRepository.selectedPet(from: userInfo)
-        self.selectedPetId = selectedPet?.petId ?? ""
-        self.userName = userInfo?.name
-        self.petName = selectedPet?.petName
+        userInfo = resolveDisplayedUserInfo()
+        selectedPet = profileRepository.selectedPet(from: userInfo)
+        selectedPetId = selectedPet?.petId ?? ""
+        userName = userInfo?.name
+        petName = selectedPet?.petName
         reloadSeasonProfileSummary()
-    }
-
-    /// 현재 실행 모드에 맞는 사용자 스냅샷을 결정합니다.
-    /// - Returns: 일반 실행에서는 저장소 원본, UI 테스트 스텁 모드에서는 in-memory 반려견 스냅샷을 반환합니다.
-    private func resolveDisplayedUserInfo() -> UserInfo? {
-        let fetchedUserInfo = profileRepository.fetchUserInfo()
-        guard isUITestPetManagementStubEnabled else {
-            uiTestPetManagementUserInfoOverride = nil
-            return fetchedUserInfo
-        }
-        if let uiTestPetManagementUserInfoOverride {
-            return uiTestPetManagementUserInfoOverride
-        }
-        let seededUserInfo = makeUITestPetManagementUserInfo(base: fetchedUserInfo)
-        uiTestPetManagementUserInfoOverride = seededUserInfo
-        return seededUserInfo
-    }
-
-    /// UI 테스트용 반려견 관리 스냅샷을 생성합니다.
-    /// - Parameter base: 실제 저장소에서 읽은 사용자 스냅샷입니다.
-    /// - Returns: 최소 1마리의 활성 반려견을 포함하는 테스트 전용 사용자 스냅샷입니다.
-    private func makeUITestPetManagementUserInfo(base: UserInfo?) -> UserInfo {
-        let seededPet = makeUITestPetManagementPet(
-            petId: "uitest-seeded-pet",
-            petName: "UITestSeedDog",
-            breed: "UITest Breed",
-            ageYears: 4,
-            gender: .female
-        )
-        return UserInfo(
-            id: base?.id ?? "uitest-seeded-user",
-            name: base?.name ?? "UITest User",
-            profile: base?.profile,
-            profileMessage: base?.profileMessage,
-            pet: [seededPet],
-            selectedPetId: seededPet.petId,
-            createdAt: base?.createdAt ?? Date().timeIntervalSince1970
-        )
-    }
-
-    /// UI 테스트용 반려견 스냅샷을 생성합니다.
-    /// - Parameters:
-    ///   - petId: 반려견 식별자입니다.
-    ///   - petName: 화면에 노출할 반려견 이름입니다.
-    ///   - breed: 화면에 노출할 견종 문자열입니다.
-    ///   - ageYears: 화면에 노출할 나이 값입니다.
-    ///   - gender: 화면에 노출할 성별 값입니다.
-    /// - Returns: 반려견 관리 UI 테스트에서 사용하는 활성 반려견 스냅샷입니다.
-    private func makeUITestPetManagementPet(
-        petId: String,
-        petName: String,
-        breed: String,
-        ageYears: Int,
-        gender: PetGender
-    ) -> PetInfo {
-        PetInfo(
-            petId: petId,
-            petName: petName,
-            petProfile: nil,
-            breed: breed,
-            ageYears: ageYears,
-            gender: gender,
-            caricatureURL: nil,
-            caricatureStatus: .ready,
-            caricatureProvider: "ui-test",
-            isActive: true
-        )
-    }
-
-    func selectPet(_ petId: String) {
-        guard activePets.contains(where: { $0.petId == petId }) else { return }
-        profileRepository.setSelectedPetId(petId, source: "setting")
-        reloadUserInfo()
-    }
-
-    func updateProfileDetails(
-        petName: String,
-        profileMessage: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender
-    ) -> Result<Void, ProfileEditValidationError> {
-        guard let current = profileRepository.fetchUserInfo() else {
-            return .failure(.userNotFound)
-        }
-
-        let validatedUserDraft = UserProfileDraft(displayName: current.name, profileMessage: profileMessage)
-        let validatedPetDraft = PetProfileDraft(
-            petName: petName,
-            breed: breed,
-            ageYearsText: ageYearsText,
-            gender: gender
-        )
-
-        let normalizedProfileMessage: String?
-        let normalizedPet: ValidatedPetProfileDraft
-        do {
-            normalizedProfileMessage = try validatedUserDraft.validated().profileMessage
-            normalizedPet = try validatedPetDraft.validated()
-        } catch let error as ProfileEditorValidationError {
-            switch error {
-            case .invalidAgeRange:
-                return .failure(.invalidAgeRange)
-            case .invalidPetName:
-                return .failure(.invalidPetName)
-            case .invalidDisplayName:
-                return .failure(.invalidDisplayName)
-            }
-        } catch {
-            return .failure(.userNotFound)
-        }
-
-        var pets = current.pet
-        let targetPetId = selectedPetId.isEmpty == false ? selectedPetId : pets.first(where: \.isActive)?.petId
-        if let targetPetId,
-           let index = pets.firstIndex(where: { $0.petId == targetPetId }) {
-            pets[index].petName = normalizedPet.petName
-            pets[index].breed = normalizedPet.breed
-            pets[index].ageYears = normalizedPet.ageYears
-            pets[index].gender = normalizedPet.gender
-        }
-
-        _ = profileRepository.save(
-            id: current.id,
-            name: current.name,
-            profile: current.profile,
-            profileMessage: normalizedProfileMessage,
-            pet: pets,
-            createdAt: current.createdAt,
-            selectedPetId: targetPetId
-        )
-
-        if let targetPetId {
-            profileRepository.setSelectedPetId(targetPetId, source: "profile_edit_save")
-        } else {
-            NotificationCenter.default.post(name: UserdefaultSetting.selectedPetDidChangeNotification, object: nil)
-        }
-        reloadUserInfo()
-        return .success(())
-    }
-
-    /// 프로필 편집 입력값(이름/문구/반려견 정보/선택 이미지)을 저장합니다.
-    /// - Parameters:
-    ///   - profileName: 사용자 표시 이름 입력값입니다.
-    ///   - profileMessage: 사용자 프로필 메시지 입력값입니다.
-    ///   - petName: 선택 반려견의 이름 입력값입니다.
-    ///   - breed: 선택 반려견의 견종 입력값입니다.
-    ///   - ageYearsText: 선택 반려견의 나이 입력값(문자열)입니다.
-    ///   - gender: 선택 반려견 성별 입력값입니다.
-    ///   - userProfileImage: 사용자가 새로 선택한 프로필 이미지입니다.
-    ///   - petProfileImage: 선택 반려견의 새 프로필 이미지입니다.
-    /// - Returns: 저장 성공/실패 결과입니다.
-    @MainActor
-    func updateProfileDetails(
-        profileName: String,
-        profileMessage: String,
-        petName: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender,
-        userProfileImage: UIImage?,
-        petProfileImage: UIImage?
-    ) async -> Result<Void, Error> {
-        let validatedUserDraft: ValidatedUserProfileDraft
-        let validatedPetDraft: ValidatedPetProfileDraft
-        do {
-            validatedUserDraft = try UserProfileDraft(
-                displayName: profileName,
-                profileMessage: profileMessage
-            ).validated()
-            validatedPetDraft = try PetProfileDraft(
-                petName: petName,
-                breed: breed,
-                ageYearsText: ageYearsText,
-                gender: gender
-            ).validated()
-        } catch let error as ProfileEditorValidationError {
-            switch error {
-            case .invalidAgeRange:
-                return .failure(ProfileEditValidationError.invalidAgeRange)
-            case .invalidDisplayName:
-                return .failure(ProfileEditValidationError.invalidDisplayName)
-            case .invalidPetName:
-                return .failure(ProfileEditValidationError.invalidPetName)
-            }
-        } catch {
-            return .failure(error)
-        }
-
-        guard let current = currentEditableUserInfo(
-            fallbackDisplayName: validatedUserDraft.displayName,
-            fallbackProfileMessage: validatedUserDraft.profileMessage
-        ) else {
-            return .failure(ProfileEditValidationError.userNotFound)
-        }
-
-        var pets = current.pet
-        let targetPetId = selectedPetId.isEmpty == false ? selectedPetId : pets.first(where: \.isActive)?.petId
-        let targetPetIndex = targetPetId.flatMap { petId in
-            pets.firstIndex(where: { $0.petId == petId })
-        }
-        if let targetPetIndex {
-            pets[targetPetIndex].petName = validatedPetDraft.petName
-            pets[targetPetIndex].breed = validatedPetDraft.breed
-            pets[targetPetIndex].ageYears = validatedPetDraft.ageYears
-            pets[targetPetIndex].gender = validatedPetDraft.gender
-        } else {
-            return .failure(ProfileEditValidationError.selectedPetNotFound)
-        }
-
-        var updatedUserProfileURL = current.profile
-        if shouldStubProfileSaveForUITest() {
-            return finalizeProfileSave(
-                current: current,
-                displayName: validatedUserDraft.displayName,
-                profileURL: updatedUserProfileURL,
-                profileMessage: validatedUserDraft.profileMessage,
-                pets: pets,
-                targetPetId: targetPetId
-            )
-        }
-
-        do {
-            if let userProfileImage {
-                guard let imageData = compressedJPEGData(for: userProfileImage) else {
-                    return .failure(ProfileEditValidationError.imageEncodingFailed)
-                }
-                updatedUserProfileURL = try await imageRepository.uploadUserProfileImage(
-                    data: imageData,
-                    ownerId: current.id
-                )
-            }
-
-            if let petProfileImage {
-                guard let targetPetIndex else {
-                    return .failure(ProfileEditValidationError.selectedPetNotFound)
-                }
-                guard let imageData = compressedJPEGData(for: petProfileImage) else {
-                    return .failure(ProfileEditValidationError.imageEncodingFailed)
-                }
-                let uploadedPetProfileURL = try await imageRepository.uploadPetProfileImage(
-                    data: imageData,
-                    ownerId: current.id
-                )
-                pets[targetPetIndex].petProfile = uploadedPetProfileURL
-            }
-        } catch {
-            return .failure(error)
-        }
-
-        return finalizeProfileSave(
-            current: current,
-            displayName: validatedUserDraft.displayName,
-            profileURL: updatedUserProfileURL,
-            profileMessage: validatedUserDraft.profileMessage,
-            pets: pets,
-            targetPetId: targetPetId
-        )
-    }
-
-    /// 현재 로그인 사용자의 회원탈퇴를 요청합니다.
-    /// - Returns: 탈퇴 처리 성공/실패 결과입니다.
-    @MainActor
-    func deleteAccount() async -> Result<Void, Error> {
-        isAccountDeletionInProgress = true
-        defer { isAccountDeletionInProgress = false }
-        do {
-            try await accountDeletionService.deleteCurrentAccount()
-            return .success(())
-        } catch {
-            return .failure(error)
-        }
-    }
-
-    /// 프로필 편집 화면에서 선택된 반려견의 캐리커처를 생성/재생성합니다.
-    @MainActor
-    func regenerateSelectedPetCaricature() async -> String {
-        guard featureFlags.isEnabled(.caricatureAsyncV1) else {
-            return "캐리커처 기능이 아직 비활성화되어 있어요."
-        }
-        guard AppFeatureGate.isAllowed(.aiGeneration, session: AppFeatureGate.currentSession()) else {
-            return "회원 전용 기능입니다. 로그인 후 다시 시도해주세요."
-        }
-        guard let currentUser = profileRepository.fetchUserInfo(),
-              let targetPet = profileRepository.selectedPet(from: currentUser) else {
-            return "반려견 정보를 찾을 수 없어 캐리커처를 생성할 수 없습니다."
-        }
-        guard let sourceImageURL = targetPet.petProfile ?? targetPet.caricatureURL,
-              sourceImageURL.isEmpty == false else {
-            return "선택된 반려견 사진이 없어 캐리커처를 생성할 수 없습니다."
-        }
-
-        isCaricatureGenerating = true
-        UserdefaultSetting.shared.updateFirstPetCaricature(status: .processing)
-        reloadUserInfo()
-        defer { isCaricatureGenerating = false }
-
-        do {
-            let response = try await caricatureClient.requestCaricature(
-                petId: targetPet.petId,
-                userId: currentUser.id,
-                sourceImageURL: sourceImageURL,
-                requestId: UUID().uuidString.lowercased()
-            )
-            guard let caricatureURL = response.caricatureURL,
-                  caricatureURL.isEmpty == false else {
-                throw CaricatureEdgeClient.RequestError.invalidResponse
-            }
-            UserdefaultSetting.shared.updateFirstPetCaricature(
-                status: .ready,
-                caricatureURL: caricatureURL,
-                provider: response.provider
-            )
-            reloadUserInfo()
-            metricTracker.track(
-                .caricatureSuccess,
-                userKey: currentUser.id,
-                featureKey: .caricatureAsyncV1,
-                payload: ["provider": response.provider ?? "unknown"]
-            )
-            return "캐리커처 생성이 완료되어 프로필에 반영됐어요."
-        } catch {
-            let message = (error as? LocalizedError)?.errorDescription ?? error.localizedDescription
-            UserdefaultSetting.shared.updateFirstPetCaricature(status: .failed)
-            reloadUserInfo()
-            metricTracker.track(
-                .caricatureFailed,
-                userKey: currentUser.id,
-                featureKey: .caricatureAsyncV1,
-                payload: ["error": message]
-            )
-            return "캐리커처 생성에 실패했습니다: \(message)"
-        }
-    }
-
-    /// 설정 화면에서 표시할 반려견 상세 요약 문구를 생성합니다.
-    /// - Parameter pet: 요약 문자열을 만들 반려견 정보입니다.
-    /// - Returns: 견종, 나이, 성별, 활성 상태를 결합한 설명 문자열입니다.
-    func petDetailsText(for pet: PetInfo) -> String {
-        let breed = pet.breed.flatMap { $0.isEmpty ? nil : $0 } ?? "견종(선택) 미입력"
-        let age = pet.ageYears.map { "\($0)세" } ?? "나이 미입력"
-        let gender = pet.gender.title
-        let activity = pet.isActive ? "활성" : "비활성"
-        return "\(breed) · \(age) · \(gender) · \(activity)"
-    }
-
-    /// 새 반려견을 추가하고 즉시 화면 상태를 갱신합니다.
-    /// - Parameters:
-    ///   - petName: 새 반려견 이름 입력값입니다.
-    ///   - breed: 새 반려견 견종 입력값입니다.
-    ///   - ageYearsText: 새 반려견 나이 입력값입니다.
-    ///   - gender: 새 반려견 성별 입력값입니다.
-    ///   - petProfileImage: 새 반려견 프로필 이미지입니다.
-    /// - Returns: 저장 성공/실패 결과입니다.
-    @MainActor
-    func addPet(
-        petName: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender,
-        petProfileImage: UIImage?
-    ) async -> Result<Void, Error> {
-        if isUITestPetManagementStubEnabled {
-            return addUITestPetStub(
-                petName: petName,
-                breed: breed,
-                ageYearsText: ageYearsText,
-                gender: gender
-            )
-        }
-        guard let current = currentEditableUserInfo(
-            fallbackDisplayName: userInfo?.name ?? "산책꾼",
-            fallbackProfileMessage: userInfo?.profileMessage
-        ) else {
-            return .failure(ProfileEditValidationError.userNotFound)
-        }
-        let imageData: Data?
-        if let petProfileImage {
-            guard let encoded = compressedJPEGData(for: petProfileImage) else {
-                return .failure(ProfileEditValidationError.imageEncodingFailed)
-            }
-            imageData = encoded
-        } else {
-            imageData = nil
-        }
-
-        do {
-            _ = try await petManagementService.addPet(
-                draft: PetProfileDraft(
-                    petName: petName,
-                    breed: breed,
-                    ageYearsText: ageYearsText,
-                    gender: gender
-                ),
-                imageData: imageData,
-                currentUser: current
-            )
-            reloadUserInfo()
-            return .success(())
-        } catch {
-            return .failure(error)
-        }
-    }
-
-    /// 기존 반려견 정보를 수정하고 즉시 화면 상태를 갱신합니다.
-    /// - Parameters:
-    ///   - petId: 수정 대상 반려견 식별자입니다.
-    ///   - petName: 반려견 이름 입력값입니다.
-    ///   - breed: 반려견 견종 입력값입니다.
-    ///   - ageYearsText: 반려견 나이 입력값입니다.
-    ///   - gender: 반려견 성별 입력값입니다.
-    ///   - petProfileImage: 새로 선택한 반려견 프로필 이미지입니다.
-    ///   - removeProfileImage: 기존 원격 이미지를 제거할지 여부입니다.
-    /// - Returns: 저장 성공/실패 결과입니다.
-    @MainActor
-    func updatePet(
-        petId: String,
-        petName: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender,
-        petProfileImage: UIImage?,
-        removeProfileImage: Bool
-    ) async -> Result<Void, Error> {
-        if isUITestPetManagementStubEnabled {
-            return updateUITestPetStub(
-                petId: petId,
-                petName: petName,
-                breed: breed,
-                ageYearsText: ageYearsText,
-                gender: gender
-            )
-        }
-        let current = userInfo ?? profileRepository.fetchUserInfo()
-        guard let current else {
-            reloadUserInfo()
-            guard let refreshed = userInfo ?? profileRepository.fetchUserInfo() else {
-                return .failure(ProfileEditValidationError.userNotFound)
-            }
-            return await updatePet(
-                current: refreshed,
-                petId: petId,
-                petName: petName,
-                breed: breed,
-                ageYearsText: ageYearsText,
-                gender: gender,
-                petProfileImage: petProfileImage,
-                removeProfileImage: removeProfileImage
-            )
-        }
-
-        return await updatePet(
-            current: current,
-            petId: petId,
-            petName: petName,
-            breed: breed,
-            ageYearsText: ageYearsText,
-            gender: gender,
-            petProfileImage: petProfileImage,
-            removeProfileImage: removeProfileImage
-        )
-    }
-
-    /// UI 테스트 전용 새 반려견 추가 결과를 in-memory 상태에 반영합니다.
-    /// - Parameters:
-    ///   - petName: 새 반려견 이름 입력값입니다.
-    ///   - breed: 새 반려견 견종 입력값입니다.
-    ///   - ageYearsText: 새 반려견 나이 입력값입니다.
-    ///   - gender: 새 반려견 성별 입력값입니다.
-    /// - Returns: 입력 검증 및 상태 반영 결과입니다.
-    private func addUITestPetStub(
-        petName: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender
-    ) -> Result<Void, Error> {
-        do {
-            let normalizedPetName = try validateRequiredName(petName, error: ProfileEditValidationError.invalidPetName)
-            let ageYears = try parseAgeYears(from: ageYearsText)
-            var current = uiTestPetManagementUserInfoOverride ?? makeUITestPetManagementUserInfo(base: profileRepository.fetchUserInfo())
-            let newPet = makeUITestPetManagementPet(
-                petId: UUID().uuidString.lowercased(),
-                petName: normalizedPetName,
-                breed: breed.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty ? "UITest Breed" : breed.trimmingCharacters(in: .whitespacesAndNewlines),
-                ageYears: ageYears ?? 1,
-                gender: gender
-            )
-            current = UserInfo(
-                id: current.id,
-                name: current.name,
-                profile: current.profile,
-                profileMessage: current.profileMessage,
-                pet: current.pet + [newPet],
-                selectedPetId: newPet.petId,
-                createdAt: current.createdAt
-            )
-            uiTestPetManagementUserInfoOverride = current
-            reloadUserInfo()
-            return .success(())
-        } catch {
-            return .failure(error)
-        }
-    }
-
-    /// UI 테스트 전용 기존 반려견 수정 결과를 in-memory 상태에 반영합니다.
-    /// - Parameters:
-    ///   - petId: 수정 대상 반려견 식별자입니다.
-    ///   - petName: 반려견 이름 입력값입니다.
-    ///   - breed: 반려견 견종 입력값입니다.
-    ///   - ageYearsText: 반려견 나이 입력값입니다.
-    ///   - gender: 반려견 성별 입력값입니다.
-    /// - Returns: 입력 검증 및 상태 반영 결과입니다.
-    private func updateUITestPetStub(
-        petId: String,
-        petName: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender
-    ) -> Result<Void, Error> {
-        do {
-            let normalizedPetName = try validateRequiredName(petName, error: ProfileEditValidationError.invalidPetName)
-            let ageYears = try parseAgeYears(from: ageYearsText)
-            var current = uiTestPetManagementUserInfoOverride ?? makeUITestPetManagementUserInfo(base: profileRepository.fetchUserInfo())
-            guard current.pet.contains(where: { $0.petId == petId }) else {
-                return .failure(ProfileEditValidationError.selectedPetNotFound)
-            }
-            current = UserInfo(
-                id: current.id,
-                name: current.name,
-                profile: current.profile,
-                profileMessage: current.profileMessage,
-                pet: current.pet.map { pet in
-                    guard pet.petId == petId else { return pet }
-                    return PetInfo(
-                        petId: pet.petId,
-                        petName: normalizedPetName,
-                        petProfile: pet.petProfile,
-                        breed: breed.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty ? nil : breed.trimmingCharacters(in: .whitespacesAndNewlines),
-                        ageYears: ageYears,
-                        gender: gender,
-                        caricatureURL: pet.caricatureURL,
-                        caricatureStatus: pet.caricatureStatus,
-                        caricatureProvider: pet.caricatureProvider,
-                        isActive: pet.isActive
-                    )
-                },
-                selectedPetId: petId,
-                createdAt: current.createdAt
-            )
-            uiTestPetManagementUserInfoOverride = current
-            reloadUserInfo()
-            return .success(())
-        } catch {
-            return .failure(error)
-        }
-    }
-
-    /// 대표 반려견을 변경하고 즉시 화면 상태를 갱신합니다.
-    /// - Parameter petId: 대표로 지정할 반려견 식별자입니다.
-    func setPrimaryPet(_ petId: String) throws {
-        guard let current = userInfo else {
-            throw ProfileEditValidationError.userNotFound
-        }
-        _ = try petManagementService.setPrimaryPet(petId: petId, currentUser: current)
-        reloadUserInfo()
-    }
-
-    /// 반려견 활성 상태를 변경하고 즉시 화면 상태를 갱신합니다.
-    /// - Parameters:
-    ///   - petId: 활성 상태를 변경할 반려견 식별자입니다.
-    ///   - isActive: 적용할 활성 상태입니다.
-    func setPetActive(_ petId: String, isActive: Bool) throws {
-        guard let current = userInfo else {
-            throw ProfileEditValidationError.userNotFound
-        }
-        do {
-            _ = try petManagementService.setPetActive(petId: petId, isActive: isActive, currentUser: current)
-            reloadUserInfo()
-        } catch let error as SettingsPetManagementError {
-            switch error {
-            case .cannotDeactivateLastActivePet:
-                throw ProfileEditValidationError.cannotDeactivateLastActivePet
-            case .imageEncodingFailed:
-                throw ProfileEditValidationError.imageEncodingFailed
-            case .petNotFound:
-                throw ProfileEditValidationError.selectedPetNotFound
-            case .userNotFound:
-                throw ProfileEditValidationError.userNotFound
-            }
-        }
-    }
-
-    /// 업로드 전 프로필 이미지를 JPEG 데이터로 압축합니다.
-    /// - Parameter image: 서버 업로드 대상으로 선택된 원본 이미지입니다.
-    /// - Returns: 인코딩 성공 시 JPEG 데이터, 실패 시 `nil`입니다.
-    private func compressedJPEGData(for image: UIImage) -> Data? {
-        image.jpegData(compressionQuality: 0.35)
-    }
-
-    /// 기존 반려견 수정 요청을 서비스 계층에 위임하고 저장 결과를 현재 화면 상태에 반영합니다.
-    /// - Parameters:
-    ///   - current: 실제 저장 기준으로 사용할 현재 사용자 스냅샷입니다.
-    ///   - petId: 수정 대상 반려견 식별자입니다.
-    ///   - petName: 반려견 이름 입력값입니다.
-    ///   - breed: 반려견 견종 입력값입니다.
-    ///   - ageYearsText: 반려견 나이 입력값입니다.
-    ///   - gender: 반려견 성별 입력값입니다.
-    ///   - petProfileImage: 새로 선택한 반려견 프로필 이미지입니다.
-    ///   - removeProfileImage: 기존 원격 이미지를 제거할지 여부입니다.
-    /// - Returns: 저장 성공/실패 결과입니다.
-    @MainActor
-    private func updatePet(
-        current: UserInfo,
-        petId: String,
-        petName: String,
-        breed: String,
-        ageYearsText: String,
-        gender: PetGender,
-        petProfileImage: UIImage?,
-        removeProfileImage: Bool
-    ) async -> Result<Void, Error> {
-        let imageData: Data?
-        if let petProfileImage {
-            guard let encoded = compressedJPEGData(for: petProfileImage) else {
-                return .failure(ProfileEditValidationError.imageEncodingFailed)
-            }
-            imageData = encoded
-        } else {
-            imageData = nil
-        }
-
-        do {
-            _ = try await petManagementService.updatePet(
-                petId: petId,
-                draft: PetProfileDraft(
-                    petName: petName,
-                    breed: breed,
-                    ageYearsText: ageYearsText,
-                    gender: gender
-                ),
-                imageData: imageData,
-                removeProfileImage: removeProfileImage,
-                currentUser: current
-            )
-            reloadUserInfo()
-            return .success(())
-        } catch let error as SettingsPetManagementError {
-            switch error {
-            case .cannotDeactivateLastActivePet:
-                return .failure(ProfileEditValidationError.cannotDeactivateLastActivePet)
-            case .imageEncodingFailed:
-                return .failure(ProfileEditValidationError.imageEncodingFailed)
-            case .petNotFound:
-                return .failure(ProfileEditValidationError.selectedPetNotFound)
-            case .userNotFound:
-                return .failure(ProfileEditValidationError.userNotFound)
-            }
-        } catch {
-            return .failure(error)
-        }
-    }
-
-    /// 필수 이름 입력값을 검증하고 trim된 문자열을 반환합니다.
-    /// - Parameters:
-    ///   - rawValue: 검증할 원본 문자열입니다.
-    ///   - error: 입력값이 비어 있을 때 반환할 에러입니다.
-    /// - Returns: 앞뒤 공백이 제거된 이름 문자열입니다.
-    private func validateRequiredName(
-        _ rawValue: String,
-        error: ProfileEditValidationError
-    ) throws -> String {
-        let normalizedValue = rawValue.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard normalizedValue.isEmpty == false else {
-            throw error
-        }
-        return normalizedValue
-    }
-
-    /// 나이 입력 문자열을 `Int?`로 해석하고 허용 범위를 검증합니다.
-    /// - Parameter value: 사용자가 입력한 나이 문자열입니다.
-    /// - Returns: 빈 문자열이면 `nil`, 숫자면 0~30 범위의 정수 값을 반환합니다.
-    private func parseAgeYears(from value: String) throws -> Int? {
-        let normalizedValue = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard normalizedValue.isEmpty == false else {
-            return nil
-        }
-        guard let ageYears = Int(normalizedValue), (0...30).contains(ageYears) else {
-            throw ProfileEditValidationError.invalidAgeRange
-        }
-        return ageYears
-    }
-
-    /// 저장 시점에 편집 가능한 사용자 스냅샷을 조회하고, 누락 시 인증 세션 기반으로 최소 스냅샷을 복구합니다.
-    /// - Parameters:
-    ///   - fallbackDisplayName: 로컬 스냅샷이 없을 때 사용할 표시 이름입니다.
-    ///   - fallbackProfileMessage: 로컬 스냅샷이 없을 때 사용할 프로필 메시지입니다.
-    /// - Returns: 저장 가능한 사용자 스냅샷이며, 복구 불가 시 `nil`입니다.
-    private func currentEditableUserInfo(
-        fallbackDisplayName: String,
-        fallbackProfileMessage: String?
-    ) -> UserInfo? {
-        if let current = profileRepository.fetchUserInfo() {
-            return current
-        }
-        reloadUserInfo()
-        if let current = profileRepository.fetchUserInfo() {
-            return current
-        }
-        guard let identity = authSessionStore.currentIdentity(),
-              identity.userId.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false else {
-            return nil
-        }
-
-        let selectedPetSnapshot = selectedPet ?? profileRepository.selectedPet(from: userInfo)
-        let recoveredPets: [PetInfo]
-        if let selectedPetSnapshot {
-            recoveredPets = [selectedPetSnapshot]
-        } else {
-            recoveredPets = [PetInfo(petName: "강아지", petProfile: nil, isActive: true)]
-        }
-        let recoveredSelectedPetId = selectedPetId.isEmpty == false
-        ? selectedPetId
-        : recoveredPets.first?.petId
-        return UserInfo(
-            id: identity.userId,
-            name: fallbackDisplayName,
-            profile: nil,
-            profileMessage: fallbackProfileMessage,
-            pet: recoveredPets,
-            selectedPetId: recoveredSelectedPetId,
-            createdAt: Date().timeIntervalSince1970
-        )
-    }
-
-    private func bindSelectedPetSync() {
-        NotificationCenter.default.publisher(for: UserdefaultSetting.selectedPetDidChangeNotification)
-            .receive(on: RunLoop.main)
-            .sink { [weak self] _ in
-                self?.reloadUserInfo()
-            }
-            .store(in: &cancellables)
-    }
-
-    /// 인증 세션 변경 알림을 구독해 설정 화면 상태를 현재 세션과 즉시 동기화합니다.
-    private func bindAuthSessionSync() {
-        NotificationCenter.default.publisher(for: .authSessionDidChange)
-            .receive(on: RunLoop.main)
-            .sink { [weak self] _ in
-                self?.handleAuthSessionDidChange()
-            }
-            .store(in: &cancellables)
-    }
-
-    /// 세션 유효성에 따라 설정 화면 캐시를 갱신/정리합니다.
-    private func handleAuthSessionDidChange() {
-        guard authSessionStore.currentTokenSession() != nil else {
-            userInfo = nil
-            selectedPet = nil
-            selectedPetId = ""
-            seasonProfileSummary = nil
-            return
-        }
-        reloadUserInfo()
-    }
-
-    private func reloadSeasonProfileSummary() {
-        struct StoredSeasonState: Decodable {
-            let weekKey: String
-            let score: Double
-            let contributionCount: Int
-        }
-
-        guard let data = UserDefaults.standard.data(forKey: "season.motion.current.v1"),
-              let decoded = try? JSONDecoder().decode(StoredSeasonState.self, from: data) else {
-            seasonProfileSummary = nil
-            return
-        }
-        let score = Int(decoded.score.rounded())
-        let rankTier: SeasonRankTier
-        if decoded.score >= SeasonRankTier.platinum.minimumScore {
-            rankTier = .platinum
-        } else if decoded.score >= SeasonRankTier.gold.minimumScore {
-            rankTier = .gold
-        } else if decoded.score >= SeasonRankTier.silver.minimumScore {
-            rankTier = .silver
-        } else if decoded.score >= SeasonRankTier.bronze.minimumScore {
-            rankTier = .bronze
-        } else {
-            rankTier = .rookie
-        }
-        seasonProfileSummary = SeasonProfileSummary(
-            weekKey: decoded.weekKey,
-            score: score,
-            rankTier: rankTier,
-            contributionCount: decoded.contributionCount
-        )
-    }
-
-    /// UI 테스트에서 프로필 저장을 원격 업로드 없이 로컬 스텁으로 처리할지 여부를 반환합니다.
-    /// - Returns: UI 테스트 전용 저장 스텁을 사용해야 하면 `true`입니다.
-    private func shouldStubProfileSaveForUITest() -> Bool {
-        ProcessInfo.processInfo.arguments.contains("-UITest.ProfileSaveStubSuccess")
-    }
-
-    /// 검증을 통과한 프로필 편집 결과를 로컬 저장소와 현재 화면 상태에 반영합니다.
-    /// - Parameters:
-    ///   - current: 저장 기준이 되는 현재 사용자 정보입니다.
-    ///   - displayName: 저장할 사용자 표시 이름입니다.
-    ///   - profileURL: 저장할 사용자 프로필 이미지 URL입니다.
-    ///   - profileMessage: 저장할 사용자 프로필 메시지입니다.
-    ///   - pets: 저장할 반려견 목록입니다.
-    ///   - targetPetId: 저장 후 선택 상태로 유지할 반려견 식별자입니다.
-    /// - Returns: 로컬 저장 반영 결과입니다.
-    private func finalizeProfileSave(
-        current: UserInfo,
-        displayName: String,
-        profileURL: String?,
-        profileMessage: String?,
-        pets: [PetInfo],
-        targetPetId: String?
-    ) -> Result<Void, Error> {
-        _ = profileRepository.save(
-            id: current.id,
-            name: displayName,
-            profile: profileURL,
-            profileMessage: profileMessage,
-            pet: pets,
-            createdAt: current.createdAt,
-            selectedPetId: targetPetId
-        )
-        if let targetPetId {
-            profileRepository.setSelectedPetId(targetPetId, source: "profile_edit_save")
-        } else {
-            NotificationCenter.default.post(name: UserdefaultSetting.selectedPetDidChangeNotification, object: nil)
-        }
-        reloadUserInfo()
-        return .success(())
     }
 }
